Read.me
Test a Perceptual Phenomenon in a Stroop task, participants are presented with a list of words, with each word displayed in a color of ink. The participant’s task is to recognize the color of the ink in which the word is printed. The task has two conditions: a congruent words condition, and an incongruent words condition.